Cambios en Controllers y Models, mayor seguridad

+++ src/controllers/administratives.js
@@ -23,7 +23,44 @@ export class AdministrativesController {
     res.json(administratives)
   }
 
+  getAdministrativeByDNI = async (req, res) => {
+    const { dni } = req.params
+
+    const administratives = await AdministrativesModel.getAdministrativeByDNI({ dni })
+
+    if (!administratives) {
+      return res.status(400).json({ status: 'Error', message: 'No existe administrativo para mostrar' })
+    }
+    res.json(administratives)
+  }
+
+  getAdministrativeByEmail = async (req, res) => {
+    const { email } = req.params
+
+    const administratives = await AdministrativesModel.getAdministrativeByEmail({ email })
+
+    if (!administratives) {
+      return res.status(400).json({ status: 'Error', message: 'No existe administrativo para mostrar' })
+    }
+    res.json(administratives)
+  }
+
+  getAdministrativeByTelefono = async (req, res) => {
+    const { telefono } = req.params
+
+    const administratives = await AdministrativesModel.getAdministrativeByTelefono({ telefono })
+
+    if (!administratives) {
+      return res.status(400).json({ status: 'Error', message: 'No existe administrativo para mostrar' })
+    }
+    res.json(administratives)
+  }
+
   createAdministrative = async (req, res) => {
+    if (!req.body.contractId) {
+      req.body.contractId = 'NULL'
+    }
+
     const validate = validateAdministrative(req.body)
 
     if (!validate.success) {
@@ -31,15 +68,40 @@ export class AdministrativesController {
       // return res.status(400).json({ status: 'Error', message: 'Error Administrative Schema' })
     }
 
-    const validateContractId = await ContractsModel.getContractById({ id: validate.data.contractId })
+    // Comprobamos si existe administrativo por DNI, email y telefono
+    const existeAdministrativoDNI = await AdministrativesModel.getAdministrativeByDNI({ dni: validate.data.dni })
 
-    if (!validateContractId) {
-      return res.status(400).json({ status: 'Error', message: `El administrativo ${validate.data.nombre} no ha sido creado correctamente por que el contractId no es correcto ` })
+    if (existeAdministrativoDNI) {
+      return res.status(400).json({ status: 'Error', message: 'Este administrativo ya exisite' })
+    }
+
+    const existeAdministrativoEmail = await AdministrativesModel.getAdministrativeByEmail({ email: validate.data.email })
+
+    if (existeAdministrativoEmail) {
+      return res.status(400).json({ status: 'Error', message: 'Este administrativo ya exisite' })
+    }
+
+    const existeAdministrativoTelefono = await AdministrativesModel.getAdministrativeByTelefono({ telefono: validate.data.telefono })
+
+    if (existeAdministrativoTelefono) {
+      return res.status(400).json({ status: 'Error', message: 'Este administrativo ya exisite' })
+    }
+
+    // Comprobamos que cuando creamos un administrativo sin contrato no pete
+
+    if (validate.data.contractId !== 'NULL') {
+      const validateContractId = await ContractsModel.getContractById({ id: validate.data.contractId })
+
+      if (!validateContractId) {
+        return res.status(400).json({ status: 'Error', message: `El administrativo ${validate.data.nombre} no ha sido creado correctamente por que el contractId no es correcto ` })
+      }
     }
 
     const newAdministrative = await AdministrativesModel.createAdministrative({ input: validate.data })
 
-    if (newAdministrative) {
+    if (newAdministrative === 1) {
+      return res.status(400).json({ status: 'Error', message: `El administrativo ${validate.data.nombre} no ha sido creado correctamente por que el contrato ya está en uso` })
+    } else if (newAdministrative) {
       return res.status(201).json({ status: 'ok', message: `Administrative ${newAdministrative.nombre} con id ${newAdministrative.employeeId} creado correctamente`, redirect: '/administratives' })
     } else {
       return res.status(400).json({ status: 'Error', message: `El administrativo ${validate.data.nombre} no ha sido creado correctamente ` })
@@ -72,6 +134,12 @@ export class AdministrativesController {
 
   deleteAdministrative = async (req, res) => {
     const { id } = req.params
+
+    const administratives = await AdministrativesModel.getAdministrativeById({ id })
+
+    if (!administratives) {
+      return res.status(400).json({ status: 'Error', message: 'No existe administrativo para eliminar' })
+    }
 
     const deletedAdministrative = await AdministrativesModel.deleteAdministrative({ id })
